Update gtm-example.md
Updated Google Ads Vendor ID and purpose consents: This is correct for Perf. Marketing according to Laura and we will not worry about Media, since Media does not have a single Google Ads Tags in GTM. SO from a GTM perspective we are fine. We need to block only Google Ads - Perf. Marketing side.